failed attempt

+++ Simulation/src/Algorithms/Algo1.java
@@ -17,10 +17,10 @@ public class Algo1 extends NavigationAlgorithm{
 		this.gui = gui;
 	}
 
-	public TreeMap<Double, ArrayList<Integer>> rotate()
+	public TreeMap<Double, ArrayList<Double>> rotate()
 	{
 		Quadcopter quad = gui.getQuad();
-		TreeMap<Double, ArrayList<Integer>> map = new TreeMap<Double, ArrayList<Integer>>();
+		TreeMap<Double, ArrayList<Double>> map = new TreeMap<Double, ArrayList<Double>>();
 		for(int i=0 ; i<36 ; i++)
 		{
 			gui.turn(10);
@@ -38,41 +38,41 @@ public class Algo1 extends NavigationAlgorithm{
 				map.get(frontDist).add(quad.getAngle());
 			}
 			else {
-				map.put(frontDist, new ArrayList<Integer>());
+				map.put(frontDist, new ArrayList<Double>());
 				map.get(frontDist).add(quad.getAngle());
 			}
 			if(map.containsKey(rightDist)) {
 				map.get(rightDist).add(quad.getAngle()+quad.getRightAngle());
 			}
 			else {
-				map.put(rightDist, new ArrayList<Integer>());
+				map.put(rightDist, new ArrayList<Double>());
 				map.get(rightDist).add(quad.getAngle()+quad.getRightAngle());
 			}
 			if(map.containsKey(leftDist)) {
 				map.get(leftDist).add(quad.getAngle()+quad.getLeftAngle());
 			}
 			else {
-				map.put(leftDist, new ArrayList<Integer>());
+				map.put(leftDist, new ArrayList<Double>());
 				map.get(leftDist).add(quad.getAngle()+quad.getLeftAngle());
 			}
 		}
 		return map;
 	}
 	public void navigate() {
-		int longestWay=0;
+		double longestWay=0;
 		double longestDistance =0;
 		boolean rotate = true; 
 		while(gui.getTime()<=600) {
 			double frontDist = gui.getQuad().getFrontLidar().getCurrentDist();
 			double rightDist = gui.getQuad().getRightLidar().getCurrentDist();
 			double leftDist = gui.getQuad().getLeftLidar().getCurrentDist();
-			Tuple<Integer,Double> tup = getLongestWayAndDistance(frontDist, rightDist, leftDist);
+			Tuple<Double,Double> tup = getLongestWayAndDistance(frontDist, rightDist, leftDist);
 			longestDistance = tup.y;
 			longestWay = tup.x;
 			
 		
 			if(rotate) {
-				TreeMap<Double, ArrayList<Integer>> map = rotate();
+				TreeMap<Double, ArrayList<Double>> map = rotate();
 				longestDistance = map.lastKey();
 				System.out.println("this " + map.lastKey());
 				longestWay = getLongest(map);
@@ -100,16 +100,16 @@ public class Algo1 extends NavigationAlgorithm{
 		}
 	}
 
-	private Tuple<Integer,Double> getLongestWayAndDistance(double frontDist, double rightDist, double leftDist) {
-		if(frontDist>=rightDist&&frontDist>=leftDist) return new Tuple<Integer, Double>(gui.getQuad().getAngle(),frontDist);
-		else if(rightDist>=leftDist) return new Tuple<Integer, Double>(gui.getQuad().getAngle()+gui.getQuad().getRightAngle(),rightDist);
-		return new Tuple<Integer, Double>(gui.getQuad().getAngle()+gui.getQuad().getLeftAngle(), leftDist);
+	private Tuple<Double,Double> getLongestWayAndDistance(double frontDist, double rightDist, double leftDist) {
+		if(frontDist>=rightDist&&frontDist>=leftDist) return new Tuple<Double, Double>(gui.getQuad().getAngle(),frontDist);
+		else if(rightDist>=leftDist) return new Tuple<Double, Double>(gui.getQuad().getAngle()+gui.getQuad().getRightAngle(),rightDist);
+		return new Tuple<Double, Double>(gui.getQuad().getAngle()+gui.getQuad().getLeftAngle(), leftDist);
 	}
 
 
 
-	private int getLongest(TreeMap<Double, ArrayList<Integer>> map) {	
-		int angle = map.get(map.lastKey()).get((int)Math.random()*(map.size()-1));
+	private double getLongest(TreeMap<Double, ArrayList<Double>> map) {	
+		double angle = map.get(map.lastKey()).get((int)Math.random()*(map.size()-1));
 		int i=0;
 		while(angle<gui.getQuad().getAngle()+100&&angle>gui.getQuad().getAngle()-100&&i<12) {
 			angle = map.get(map.lastKey()).get((int)Math.random()*(map.size()-1));
